+++ EventsWidget/EventsTimelineEntry.swift
@@ -1,0 +1,8 @@
+import WidgetKit
+
+struct EventsTimelineEntry: TimelineEntry {
+    
+    let date: Date
+    let configuration: ConfigurationIntent
+    
+}
